requirements.txt
affine==2.3.0
alembic==1.3.0
argh==0.26.2
attrs==19.3.0
Babel==2.7.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
decorator==4.4.1
earthpy==0.7.5
Fiona==1.8.11
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Table==0.5.0
Flask-WTF==0.14.2
geopandas==0.6.1
google-api-python-client==1.7.11
google-auth==1.7.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gunicorn==20.0.4
heroku==0.1.4
httplib2==0.14.0
idna==2.8
imageio==2.6.1
imbalanced-learn==0.4.3
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
kiwisolver==1.0.1
lightgbm==2.2.3
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mpld3==0.3
munch==2.5.0
names==0.3.0
networkx==2.4
numpy==1.17.4
oauth2client==4.1.3
oauthlib==3.1.0
pandas==0.24.2
parse==1.11.1
pathtools==0.1.2
Pillow==6.2.1
playsound==1.2.2
psutil==5.6.1
psycopg2-binary==2.8.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
PyDrive==1.3.1
pyparsing==2.3.1
pyproj==2.4.1
python-dateutil==2.8.1
python-editor==1.0.4
python-speech-features==0.6
pytz==2019.1
PyWavelets==1.1.1
PyYAML==5.1.2
rasterio==1.1.0
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
Rtree==0.8.3
scikit-image==0.16.2
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
Shapely==1.6.4.post2
six==1.12.0
snuggs==1.4.7
SQLAlchemy==1.3.10
uritemplate==3.0.0
urllib3==1.25.6
watchdog==0.9.0
Werkzeug==0.15.6
WTForms==2.2.1
xgboost==0.82
